train_model as a function of the vae class
