+++ pyLOM/VAE/__init__.py
@@ -8,7 +8,8 @@
 
 __VERSION__ = '2.0.1'
 
-from .wrapper import VariationalAutoencoder, VariationalDecoder, VariationalEncoder, EarlyStopper
-from .utils   import Dataset, save
+from .wrapper   import VariationalAutoencoder, VariationalDecoder, VariationalEncoder
+from .callbacks import EarlyStopper
+from .utils     import Dataset, save
 
 del wrapper
